backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-limiter==3.8.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyJWT==2.9.0
gunicorn==23.0.0
